Produce tests which have dependencies on others tests
This sort of works: if you narrow the selection of a test to just
one test, it runs all the tests prior to that one, however if you
choose a full class, it will run each test once, which each test
running all its dependencies.

It's not clear how to map the last test on to 'all'?

And that doesn't address the cases where we might match some but not
all tests and still want to get the dependencies right.