Get rid of testresources, too much complexity
There's no point managing a complex graph. Instead just keep
track of the linear sequence of dependencies.

If a test has been run in this process, don't run it again.

Some jiggery pokery is required to make sure that setUp and tearDown
are handled appropriately.

Group tests (for concurrency) on the name of the yaml file.

Thanks to @FND for some rubber ducking to get this figured out. He
helped me clear away the dismal hate for testr, subunit, etc. I
still hate them but at least they aren't in the way anymore.