Provide a $LOCATION, $NETLOC, $SCHEME variables in YAML
This makes it easier to make references to prior requests and
confirm URL construction.

If there is a Location header in a response, then the TestCase
recieves a location attribute to be referred to by subsequent
requests. This is to encourage a more RESTful approach.

In a future commit and reference to all the headers and body
text may be saved, assuming it seems useful.