Several content-decoding related bug fixes
Response content is now decoded according to the provided charset
in the content-type header. This is exercised by the request for the
front page of the the Russian Google.

Content which ought to be binary is not decoded. This is likely to be
somewhat buggy until we have some good usage patterns. The
SimpleWsgi app has been adjusted to allow a way of testing this: if
you send a request with an Accept header, the content-type in the
response is that given in the accept header.

Please note that the SimpleWsgi app is just for use by gabbi to test
itself. In real testing situations the test author provides a WSGI
app.