Implement declarable fixtures
A fixture is a sublass of GabbiFixture. It is a singleton which
is started on the first test run and then stop at the end of the
TestSuite. To manage stopping, the testsuite is now a GabbiSuite,
a subclass of the usual TestSuite with run overridden to shutdown
the fixture after the suite has completed. The normal class and module
way of doing this will not work because the test cases have
insufficient sense of the modules they come from.

Each YAML file can have multiple fixtures, started in order. They
are shutdown in reverse order.

What a fixture does is left as an exercise. Or rather, we'll see.

This is not fully done, but I'm asleep, so committing.