Delay wsgi interception until after fixtures
Some WSGI applications require a lot of prior setup before they can
be use. Since we want most of that setup to happen in fixtures,
wsgi-interception is delayed until after all extant fixtures have been
started. This is done with a built in default fixture that is run as
the most deeply nested of the context managers.

The intecept callable is passed on along to the tests so that suite
can be made aware of and use it when using the intercept context
manager. If the callable is a function, when it is passed into the
class builder, it can become bound to the TestCase. This binding is
unbound in the caller.