Correctly handle fixtures when selecting limited tests
If a pattern is passed to the test runner such that only
some tests in a YAML file are to be run, the unrequested
tests are represented in the TestSuite as empty TestSuites.
These do not have a 'fixtures' attribute. This can result
in the WSGI-intercept not being discovered and installed.

Instead of asking for the first test when looking for fixtures,
we traverse the suite until we find the first test which has
it.