Handle $RESPONSE templates in the current response
In some APIs it is useful to refer to the response of a previous
request in the response of this current request to, for example,
verify that an ID continues to be used.

Up to now $RESPONSE handling worked in url and data, this change
adds it to response_strings and response_json_paths.

As with all of the $RESPONSE handling, this is power that should
really only be used when testing existing APIs. If you need a lot of
this when you are developing new APIs YOU ARE DOING IT WRONG.