Make response handlers more pluggable
We can pass a list of ResponsHandler classes to build_tests and these
will be used to extend the built in list.

This feels like the simplest way to do it.

A test handler (and tests of it) has been added.

The test key string has been reified as self._key.