Adjust the ResponseHandler to require less of subclasses
Now the loop is done in the __call__ in the super class. The subclass
need only implement action() to evaluate one line in the data provided
to the test key that this handler handles.

A stab has been made at documenting this feature but the language
feels really stilted right now, probably because I'm tired.
Improvements welcomed.