Wrap the http request to trap WSGIAppError and reraise
To create a strong boundary between the client and the server,
wsgi-intercept wraps exceptions raised in the intercepted app in a
WSGIAppError. In most situations this is a good thing. However if
gabbi is being used to do TDD against a nascent API this behavior
can obscure useful traceback information when there is a
catastrophic and untrapped exception in the server code.

To deal with this the WSGIAppError is caught and the contained
exception is reraised in all its glory to spew into STDERR for
debugging purposes.