Ensure defaults update correctly
String default values were updating appropriately, but not lists or
dicts. A test_update method is added to handle these correctly.
The 'data' key is treated specially as it might be a string, a list
or a dict. Whatever form it is, we want to replace the default, not
merge.