Allow single or double quotes in replacers
Such that ENVIRON['foo'] and ENVIRON["foo"] will both work but
ENVIRON["foo'] will not.

In the process regex handling is refactored to allow each of the
replaces to use the same basic regex and named groups rather than
numbered. Should allow for future adjustments like this to be less
painful.

Fixes #58