Be more robust in processing content-type charset
We can't rely on there being only one parameter in the content-type,
so we need to do some processing to extract exactly the charset.

Thanks to @ravichandrasadineni for pointing out the problem and
providing a possible solution. This solution differs slightly from
his, but is the same in spirit.

The simple_wsgi app has been updated to return a content-type which
will exercise this code.

Fixes #66