Write out the error message with gabbi run
We simply stringify the args of the exception that is at err[1].
For the most part this usually just gets the error message passed to
the AssertionError or other exception. In the case of errors like
IOError the first arg is an errno, but the outcome ends up being as
desired.