Extract a COLORMAP
This is simply to make things easier to override in a subclass, for
when I don't want YELLOW or some such.

Note that this rather begs the question of how one injects a
subclass but we'll worry about that some other time. It's currently
not germane, we just want to enable to future capability.