Allow envivonment variables to set verbose color
This is starting to get a bit out of hand now, but is a useful
feature. That is, if we're going to have the general feature of
colored output, then being able to choose, from the calling harness
(often shell scripts) is useful.

Probably useful to merge this soon so that it is out there for
feedback.