Disallow "METHOD: url" in defaults
Otherwise it becomes difficult to manage on a few dimensions:

* the code to do it is complex
* the readability of the test is damaged

So now the default dict is validated. If it has uppercase keys a
GabbiFormatError is raised. A non-gabbi test for this has been
added, along with a test for duplicate METHOD keys in one test.
The existing method_shortcut.yaml tests have been updated to reflect
the change in defaults.

Some further cleanups on the tests are probably required and docs are
still missing, but this is almost working.