Extract create_url to utils
The case._parse_url method was overly complex and merged several
different pieces of functionality. This change extracts the part
that creates a fully qualified url when the url is not already fully
qualified.

It also moves query_parameter extension to its own method.

The resulting code is still complex (and has further changes
forthcoming) but is moving in the right direction.