Tune up and test query_parameters test directive
The query_parameters directive extends the query parameters on a url
defined in a test. The method to handle that was extracted in the
previous commit but was not well tested. This changed adds better
tests and identifies and fixes a misfeature:

In previous code query_parameters would only work on URLs that were
not fully qualified with a sheme of their own. Now it works on any
URL which makes more sense.