Tidy single test creation
This is an experiment to see if there is a more clear way to
understand and implement the creation of a single test in a test
suite. This may not be the best way to do it, but it might stimulate
some ideas or be enough of a step in the right direction to keep it.

The change is the creation of a TestMaker class that provides a
make_one_test method. The TestMaker instance encapsulates all the
HTTPTestCase values which are the same for each test in a single
YAML file. make_one_test accepts the ones that are different.

This is a good demonstration of the way in which much of gabbi
evolved from a process of experimentation to get the desired results
in the face of the fairly weird/complex/wrong/whatever interface of
unittest and its ecosystem. Clearly things are messy but all that
really means is that they can get better with further iterative
development.