Clean up the TestMaker concept
The concept of the TestMaker was correct but in the process of
making it less complex internally (with the addition of private
method), it became clear that it was not being called correctly.
Several variables being passed into make_one_test because they were
considered per-test were not per-test but per file so they are
now passed into __init__ instead.

There is a data-not-behavior changing change in this commit:
test['name'] is now processed to become full name of the test
as used to name the HTTPTestCase (see the change in
test_driver.py). This was done to support extracting the private
methods in TestMaker.

/cc @FND you were mostly correct about the loop that is possible
across the make_one_test function, I guess I just can't mentall
make that many changes in one go.