Update and move a comment to reflect reality
Early on the comment was true: If there was no host or intercept an
empty suite would be returned. This proved to mask errors in test
creation in large testing systems so a more visible and immediate
failure was implemented. The comment, however, did not get updated.
Now it is.