Add initial support for jsonpath sort and filter
This requires a slight modification to jsonpath_rw_ext to get it
to return a single datum in context from a sort, rather than a plain
list. Without the former it is not possible to do work the value
returned. Will try to get a fix applied there.

Docs to come.