Switch to use urllib3 from httplib2
httplib2's maintenance status is in doubt so using urllib3 is a bit
more sane because it is the active and preferred choice for http
library that provides a high level of control and low level of magic.

This changes both the client for making the requests, and the
wsgi-intercept handling.

The details of the changes are in gabbi.httpclient which creates
a lightweight facade over the urllib3.PoolManager.request method
to transform the results into something that has the shape of an
httplib2 response, content pair.