Conflate the Response and Content handler concepts.
A ResponseHandler is a class that has a 'test_key_suffix', a
ContentHandler is a class that has an 'accepts' method. A handler may
be one or the other, or both. At test build time a list of available
handlers is registered. If the class is a response handler it is
added to response_handlers list on the test class. If the class is a
content handler it is added to the separate content_handlers list on
the test class.

The build_test method is modified to accept a content_handlers argument
in addition to the response_handlers argument. Either may be used,
duplication is handled. This is done to preserve backwards
compatibility.

gabbi-run still only has the response handlers argument. I'm not sure
if adding another one is the right thing to do or not.

The ordering of how handlers are loaded has been changed so that custom
content handlers come before custom response handlers come before built
ins. This wasn't an issue with response handlers (which were indexed by
keys) but is with content handlers which do an ordered traversal to find
a class that accepts().

Thus far this "just works" but I've not yet tried anything outside the
JSON norms.