Explore an HTML content handler
Do this on a seprate branch as it is has exposed some issues with
the handling of state on the HTTPTestCase class.

Basically this creates an html_handler based on gabbi-html created
by @FND. The main features can be see in html_handler.yaml.

This was done to prove out the ContentHandler concept and it mostly
works except for the aforementioned issues with state in tests and
perhaps an issue with dumps and loads needing to date a test object
instead of just data. Otherwise there's not not enough access to
context that can be useful in the dump and loading.

This branch will fail tests.