Cleanup the handling of handlers on HTTPTestCase
After each test the base setting of the test sometimes needs to
be reset. I'm not sure this is the best way to do things, but it
does work.

I don't want to put the HTML handler in the code, but it was
useful for fleshing out the test. So the next step is to extract
the relevant changes from this branch that are not the
html_content_handler or relations, and then package that up
as its own thing.