Add support for pytest
This change adds a gabbi.driver.py_test_generator function which
can be used to generate gabbi tests that can be run by py.test.

This works except for one important but not necessarily critical
feature: If you select one test from a gabbit (using the -k arg
to py.test) the test.prior stack is traverse to the first test
and that first test is run, but the others are not. We may be able
to figure out to fix this, but for now I don't have any ideas.

I believe this change  is still useful without that feature, but
it will need to be heavily documented.

Next steps:

* documentation
* explore options for dealing with prior