Make the list of content_ and response_handlers instance oriented
This means that, if necessary, each test can have different lists.
Most of the time this won't be the case, but in environments where
tests started with different handlers are running the same process
it can get messy.

This changes the interface to test_suite_from_dict, because we need
to pass the created handler objects to the code that creates the test
objects. I had assumed that content handlers would indicate version 2
so maybe this is fine. (In any case it only impacts people who are
calling test_suite_from_dict themselves, turns out it is more common
to call build tests.)