Add documentation for ContentHandlers
In the process of writing, some slight adjustments to the interface
were made:

* ContentHandlers direcly inherit from ResponseHandler rather than
  doing a Mixin. This makes it easier to create your own.

* Rather than defining gen_replacer, a subclass defines replacer,
  which gen_replacer calls.

The documentation is challenging to get right. I suspect there are
many improvements that could be made to make it more informative
and easier to follow. I think I'm probalby too close to it. @FND
will likely have some useful suggestions.

Still todo: docstrings on the new classes.