First pass at implementing $LAST_URL feature
If $LAST_URL is used anywhere a replacer is accepted, it will
be replaced with the URL from the previous test.

At the moment the URL is the full url that is passed into the http
client. Thus it includes the scheme, the netloc and any query
parameters that may have been added by the use of the
`query_parameters` test key.

It's not certain that this is the right choice. The other option is
to use the value of `url` that is in the test input. The is what's
actually the value of the url or {GET,POST,..} key in the test.
This is problematic though as there are at least two ways this could
be interpreted:

* the literal string that is in test['url'] in the test data
  structure
* that string after replacers have been run

Using the final full URL seems to suit the use cases I can think of
and feels somehow less ambiguous so that's what I implemented, but
it too has some issues:

* If $LAST_URL has query parameters in it, trying to _replace_ (not
  extend) those parametes via `query_parameters` doesn't really work.
  There's a failing test to demonstrate this problem.

(there may be others I haven't thought of yet)

Input desired from @FND and others please.