Tune up the docs for content-handlers a bit
Thanks to comments from @FND this now has a more useful structure
that orders and separate things by topic and complexity.

* introduce the general topic
* say how to use existing handlers
* explain how to create them

This is an improvement but I'm not certain it is yet perfect.
That's okay. If it is good enough we can move forward and move
it in the direction of perfect later.