Fix failing tests
Tests in test_runner.py can intermittently fail because test_parse_url
was using case.BASE_TEST without first copying it. Therefore changes
that were made in test_parse_url could be reflected elsewhere. This
points to a generic problem in gabbi itself: There needs to be better
management of global config-like items.

test-limit.sh was failing because the test it uses to signal limiting
was moved from driver to suitemaker. The string used in a grep
had to be corrected.