Send a warning when a yaml filename has an underscore
TestSuites are grouped by test runners by the name that is
generated for the tests within. To ease that grouping it is
useful to have filenames with a known set of characters.

Grouping is also used for py.test fixture handling.

This change raises a warning if a filename has an '_' in it,
effectively declaring that though it is possible to use '_',
it can impact grouping.

Fixes #144