Treat replacers that look like numbers in data as numbers
When the value of the `data` key is not a string (is a dict or list)
each item in that structure could have a magical substitution
variable in it. In the case of JSON if the structured data has replacers
in it, those replacers are themselves strings, so the structure
gets things typed strings. To work around this we make it so things
which look like numbers are made into numbers but only in the
data key.