Further cleanups for retype_numerals
The simple_wsgi test application takes query strings and then sets
that information as json in the response body. Initially this was
setting things which look like numbers as strings which was then
conflicting with the expectations provided by replace_template.

simple_wsgi has now been updated to think of numbers as numbers
and related tests have been changed accordingly.

This is getting rather hairy but that's inevitable when we are
in a situation where types are ambiguous.