Extract sys.stdin handling from the input_files
This is probably a personal preference not a technical one, given that
we are going to be making two separate calls to run_suite, it makes more
sense to make that choice outside of a for loop, not within it. This
makes the reason for the different calls a bit more evident and means
that everything in the input_files list is of the same type and no type
checking is required.