Fix pytest warnings by using pytest_generate_tests
This is not done. Though it works to remove the warnings that
are created when using yield to generate the tests, it is cumbersome
for the tester to use: pytest_generate_tests needs to be in the test_*
module and that module needs to have test_pytest in itself in some
fashion. What would be better for the the user of gabbi (the test
author) to be able to only call a method with the right arguments
and for the pytest details to be hidden. This may be possible but
pushing this up for sharing and discussion first.

/cc @FND, @jasonamyers