Pretty print JSON bodies in verbose tests
If a request or response body is JSON, then use the loads and dumps
methods on the JSON content handler to present it in a pretty
fashion. See the TODO in gabbi/httpclient.py for future expansion
options.

This reuses a test_verbose.yaml file added in pull request #183
that didn't seem to be doing anything.

Tests for this are in test_runner.py as it provides useful stdout
capturing infrastructure.

Fixes #182