Improve coverage testing and handling
Because the unittest and pytest runners handle the gabbi test
suites and fixtures differently, testing coverage from just
the unittest side is incomplete and potentially misleading.

The changes here include:

* adding a pytest-cov tox target to do coverage from pytest
  output is in html, put into the default 'htmlcov' directory
  (the cover target's output goes into the directory 'cover')
* adding the pytest-cov module to test-requirements.txt
* adjusting both coverage tests so coverage of the tests are
  not included in the reports, just the "real" code
* adding the artifacts from these additions to .gitignore

Sadly (but goodly), the result of these changes is that gabbi's
coverage sorts are not as bigly as they should be. With the help of
these new reports we'll make gabbi's test coverage great again.