Handle exceptions when starting fixtures
Some test runners will swallow exceptions and tracebacks that happen
during the start_fixture method of a Fixture. This can lead to a
great deal of confusion when creating new fixtures. This change
ensures that the traceback is made available by forcing an error to
be reported on the first test in the test suite using the fixture.
Additional detail can be found in the comment associated with the
change.

The added test_suite file tests this change by creating a suite with
two tests that use a fixture that will raise an exception in
start_fixture. The test confirms that the right test is errored,
that both are skipped and that the traceback includes expected
information.

Fixes #199