Avoid duplicating prefix when re-using urls
If a reponse includes an absolute link, re-using that link as the
url in a subsequent test when a prefix is being used in the tests
can result in a 404 because the prefix will be duplicated.

This change ensures that the prefix will not be duplicated if the
URL already begins with the prefix.

Fixes #165