Make JSON valid after replacing RESPONSE instances for a given key
Resolves #147, having $RESPONSE replacements that contained integer
or decimal values would wind up as quoted strings after substitution.
The approach this diff takes (after substitution) is:

1. Attempt to decode the passed `string` param as a dict. If this fails,
the enclosing try/except will simply return the original string. (e.g.
we have `'hotdog'` instead of `{"hotdog": "12"}`).
2. Go through each key-value pairing in the dict (`json_rep`) and
attempt to coerce str/unicode values to float. We choose float here as
`float('10')` is also a valid parsing to an int for handing off to
`json.dumps`. If we cannot parse the value, then we simply continue to
the next pairing.
3. Stringify the result via `json.dumps`.