Proof of concept for `coerce` method
In this approach we provide a coerce method to the content handlers.
This way, after the values have been substited in, we can enforce that
they have the correct types and we can allow the user to do this per
content type.

I have not done testing with coercing replaced arrays and objects, but
extending the `coerce` in `JSONHandler` wouldn't be that troublesome by
simply adding an extra `json.loads`, I think. I will have to test this
hypothesis.

At the moment this passes `tox -epy27` and `tox -epep8`.