Fix failing error in prexisting tests
The JSONHandler response_path action was failing its tests because of
type mismatches. Since the action is happening within the context of the
content handler, it made sense to coerce the values prior to comparison
(except for when the comparison is a regex).

This required mucking around some more with the JSONHandler's coerce
method:

1. Try to parse the given value.
2. If value is a dict (prior or after parsing), coerce each of it's
fields as a new dict. (This may need recursive `coerce` calls instead of
the basic `json.loads` I have now).
3. Including or excluding step 2, dump the result as a string.

The downside with this coerce approach is that people implementing
custom handlers _must_ include custom coerce calls and if action methods
are to be implemented that will be doing comparisons, they should
probably be doing them by coercing the values (as we have here).