Patch up check in `assert_in_or_print_output` to handle dict
Sometimes with the coercions in place, expected values passed in to
testing for strings will wind up as dicts (i.e. since no content handler
is passed to `replace_template`, no coercion takes place, and a dict
remains when parsed. One approach is to allow the coercion, but this
would munge values up based on what `json.dumps` produces.

This commit checks for the dict case, and checks for un-enclosed key-val
pairs. Keys will always be strings, but the value may or may not be a
string, in which case the value might wind up quoted or not. N.B. this
could probably be done more elegantly.