Remove .testrepository/times.dbm each test run
This overcomes a long standing bug in the testr tool: it doesn't
specify a dbm type to use, using a default which changes from
one python version to another (depending on build settings). It
is also unable to identify what was already used (which would be a
normal thing to do) and use that. Instead it just fails awkwardly.

The workaround is to remove the times.db file before each run
because it is not relevant.