Add optional `test` param for custom content handlers (#212)
Add optional `test` param for custom content handlers

This resolves #211.

`multipart/form-data` transmission requires modifying the `Content-Type`
header with a `boundary` section identifier, and as such a custom
content handler would not have previously had access to the running test
case it was working on. This should hopefully be useful for other custom
content handlers in the future.